apis/requirements.external.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --output-file=requirements.external.txt ../gristcli/pyproject.toml ../models/pyproject.toml ../services/pyproject.toml requirements.external.in
#
aiofiles==24.1.0
    # via python-keycloak
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
api-entreprise==3.7.2
    # via -r requirements.external.in
async-property==0.2.2
    # via python-keycloak
authlib==1.6.1
    # via -r requirements.external.in
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   pyogrio
    #   pyproj
    #   requests
    #   sentry-sdk
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.3
    # via requests
click==8.2.1
    # via
    #   rich-toolkit
    #   typer
    #   uvicorn
cryptography==45.0.5
    # via
    #   authlib
    #   jwcrypto
    #   models (../models/pyproject.toml)
deprecation==2.1.0
    # via python-keycloak
dnspython==2.7.0
    # via email-validator
email-validator==2.2.0
    # via
    #   fastapi
    #   pydantic
fastapi[standard]==0.116.1
    # via -r requirements.external.in
fastapi-cli[standard]==0.0.8
    # via fastapi
fastapi-cloud-cli==0.1.5
    # via fastapi-cli
geoalchemy2==0.17.1
    # via models (../models/pyproject.toml)
geopandas==1.1.1
    # via models (../models/pyproject.toml)
greenlet==3.2.4
    # via sqlalchemy
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   fastapi
    #   fastapi-cloud-cli
    #   python-keycloak
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
jinja2==3.1.6
    # via fastapi
jwcrypto==1.5.6
    # via python-keycloak
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
marshmallow==3.26.1
    # via
    #   api-entreprise
    #   marshmallow-dataclass
    #   marshmallow-jsonschema
    #   marshmallow-sqlalchemy
marshmallow-dataclass==8.7.1
    # via
    #   api-entreprise
    #   gristcli (../gristcli/pyproject.toml)
    #   models (../models/pyproject.toml)
marshmallow-jsonschema==0.13.0
    # via
    #   gristcli (../gristcli/pyproject.toml)
    #   models (../models/pyproject.toml)
marshmallow-sqlalchemy==1.4.2
    # via
    #   gristcli (../gristcli/pyproject.toml)
    #   models (../models/pyproject.toml)
mdurl==0.1.2
    # via markdown-it-py
mypy-extensions==1.1.0
    # via typing-inspect
numpy==2.3.2
    # via
    #   geopandas
    #   pandas
    #   pyogrio
    #   shapely
packaging==25.0
    # via
    #   deprecation
    #   geoalchemy2
    #   geopandas
    #   marshmallow
    #   pyogrio
pandas==2.3.2
    # via
    #   geopandas
    #   gristcli (../gristcli/pyproject.toml)
    #   models (../models/pyproject.toml)
prometheus-client==0.22.1
    # via prometheus-fastapi-instrumentator
prometheus-fastapi-instrumentator==7.1.0
    # via -r requirements.external.in
psycopg[pool]==3.2.9
    # via -r requirements.external.in
psycopg-binary==3.2.9
    # via -r requirements.external.in
psycopg-pool==3.2.6
    # via psycopg
pycparser==2.22
    # via cffi
pydantic[email]==2.11.7
    # via
    #   fastapi
    #   fastapi-cloud-cli
pydantic-core==2.33.2
    # via pydantic
pygments==2.19.2
    # via rich
pyogrio==0.11.1
    # via geopandas
pyproj==3.7.2
    # via geopandas
pyrate-limiter==2.10.0
    # via api-entreprise
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.1
    # via uvicorn
python-keycloak==5.8.1
    # via -r requirements.external.in
python-multipart==0.0.20
    # via fastapi
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via uvicorn
redis==4.6.0
    # via api-entreprise
requests==2.32.5
    # via
    #   -r requirements.external.in
    #   api-entreprise
    #   python-keycloak
    #   requests-toolbelt
requests-toolbelt==1.0.0
    # via python-keycloak
rich==14.1.0
    # via
    #   rich-toolkit
    #   typer
rich-toolkit==0.15.0
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
rignore==0.6.4
    # via fastapi-cloud-cli
sentry-sdk==2.35.0
    # via fastapi-cloud-cli
shapely==2.1.1
    # via geopandas
shellingham==1.5.4
    # via typer
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.43
    # via
    #   -r requirements.external.in
    #   geoalchemy2
    #   gristcli (../gristcli/pyproject.toml)
    #   marshmallow-sqlalchemy
    #   models (../models/pyproject.toml)
sqlparse==0.5.3
    # via -r requirements.external.in
starlette==0.47.2
    # via
    #   fastapi
    #   prometheus-fastapi-instrumentator
typeguard==4.4.4
    # via marshmallow-dataclass
typer==0.16.1
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
typing-extensions==4.14.1
    # via
    #   fastapi
    #   jwcrypto
    #   psycopg-pool
    #   pydantic
    #   pydantic-core
    #   rich-toolkit
    #   sqlalchemy
    #   typeguard
    #   typer
    #   typing-inspect
    #   typing-inspection
typing-inspect==0.9.0
    # via marshmallow-dataclass
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via
    #   requests
    #   sentry-sdk
uvicorn[standard]==0.35.0
    # via
    #   fastapi
    #   fastapi-cli
    #   fastapi-cloud-cli
uvloop==0.21.0
    # via uvicorn
watchfiles==1.1.0
    # via uvicorn
websockets==15.0.1
    # via uvicorn
